Extracted state to a separate class. Correct determination of the same item and content so blinking is gone
