+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/payments/cardreader/hub/CardReaderHubViewHolder.kt
@@ -15,12 +15,12 @@ private const val DISABLED_BUTTON_ALPHA = 0.5f
 
 abstract class CardReaderHubViewHolder(val parent: ViewGroup, @LayoutRes layout: Int) :
     RecyclerView.ViewHolder(LayoutInflater.from(parent.context).inflate(layout, parent, false)) {
-    abstract fun onBind(uiState: CardReaderHubViewModel.CardReaderHubViewState.ListItem)
+    abstract fun onBind(uiState: CardReaderHubViewState.ListItem)
 
     class RowViewHolder(parent: ViewGroup) : CardReaderHubViewHolder(parent, R.layout.card_reader_hub_list_item) {
         var binding: CardReaderHubListItemBinding = CardReaderHubListItemBinding.bind(itemView)
-        override fun onBind(uiState: CardReaderHubViewModel.CardReaderHubViewState.ListItem) {
-            uiState as CardReaderHubViewModel.CardReaderHubViewState.ListItem.NonToggleableListItem
+        override fun onBind(uiState: CardReaderHubViewState.ListItem) {
+            uiState as CardReaderHubViewState.ListItem.NonToggleableListItem
             binding.cardReaderHubListItemLabelTv.text = UiHelpers.getTextOfUiString(itemView.context, uiState.label)
             UiHelpers.setTextOrHide(binding.cardReaderHubListItemDescriptionTv, uiState.description)
             binding.cardReaderMenuIcon.setImageResource(uiState.icon)
@@ -47,8 +47,8 @@ abstract class CardReaderHubViewHolder(val parent: ViewGroup, @LayoutRes layout:
     class ToggleableViewHolder(parent: ViewGroup) :
         CardReaderHubViewHolder(parent, R.layout.card_reader_hub_toggelable_item) {
         var binding: CardReaderHubToggelableItemBinding = CardReaderHubToggelableItemBinding.bind(itemView)
-        override fun onBind(uiState: CardReaderHubViewModel.CardReaderHubViewState.ListItem) {
-            uiState as CardReaderHubViewModel.CardReaderHubViewState.ListItem.ToggleableListItem
+        override fun onBind(uiState: CardReaderHubViewState.ListItem) {
+            uiState as CardReaderHubViewState.ListItem.ToggleableListItem
             binding.cardReaderHubListItemLabelTv.text = UiHelpers.getTextOfUiString(itemView.context, uiState.label)
             binding.cardReaderMenuIcon.setImageResource(uiState.icon)
             UiHelpers.setTextOrHide(binding.cardReaderHubListItemDescriptionTv, uiState.description)
@@ -73,15 +73,15 @@ abstract class CardReaderHubViewHolder(val parent: ViewGroup, @LayoutRes layout:
     class HeaderViewHolder(parent: ViewGroup) :
         CardReaderHubViewHolder(parent, R.layout.card_reader_hub_header) {
         var binding: CardReaderHubHeaderBinding = CardReaderHubHeaderBinding.bind(itemView)
-        override fun onBind(uiState: CardReaderHubViewModel.CardReaderHubViewState.ListItem) {
-            uiState as CardReaderHubViewModel.CardReaderHubViewState.ListItem.HeaderItem
+        override fun onBind(uiState: CardReaderHubViewState.ListItem) {
+            uiState as CardReaderHubViewState.ListItem.HeaderItem
             binding.cardReaderHubHeaderTv.text = UiHelpers.getTextOfUiString(itemView.context, uiState.label)
         }
     }
 
     class GapBetweenSectionsViewHolder(parent: ViewGroup) :
         CardReaderHubViewHolder(parent, R.layout.card_reader_hub_gap_between_sections) {
-        override fun onBind(uiState: CardReaderHubViewModel.CardReaderHubViewState.ListItem) {
+        override fun onBind(uiState: CardReaderHubViewState.ListItem) {
             // no-op
         }
     }
